// pwa/src/app/models/user.ts
import { Group } from './group';
import { NotificationSettings } from './notification-settings';
import { PersonalSettings } from './personal-settings';
import { TelegramSettings } from './telegram-settings';
import { WorkTimeSettings } from './work-time-settings';

export interface User {
  id:                            number;
  uid:                           string;
  full_name:                     string;
  email:                         string;
  username:                      string;
  avatar_initials_url:           string;
  avatar_uploaded_url:           string;
  initials:                      string;
  avatar_type:                   number;
  lng:                           string;
  timezone:                      string;
  theme:                         string;
  created:                       Date;
  updated:                       Date;
  activated:                     boolean;
  ui_version:                    number;
  virtual:                       boolean;
  email_blocked:                 null;
  email_blocked_reason:          null;
  delete_requested_at:           null;
  telegram_id:                   null;
  telegram_settings:             TelegramSettings;
  permissions:                   number;
  own_permissions:               number;
  groups:                        Group[];
  company_id:                    number;
  user_id?:                       number;
  default_space_id:              null;
  role:                          number;
  email_frequency:               number;
  email_settings:                null;
  slack_id:                      null;
  slack_settings:                null;
  notification_settings:         NotificationSettings;
  notification_enabled_channels: string[];
  slack_private_channel_id:      null;
  telegram_sd_bot_enabled:       boolean;
  invite_last_sent_at:           Date;
  apps_permissions:              number;
  external:                      boolean;
  last_request_date:             Date;
  last_request_method:           string;
  work_time_settings:            WorkTimeSettings;
  personal_settings:             PersonalSettings;
  locked:                        boolean;
  has_password:                  boolean;
}